made the view passed to gameGUi non null, fixed warnings

+++ ios gl tutorial/GameEntity.m
@@ -17,9 +17,9 @@
     return self;
 }
 -(void)loadToBuffers:(const Vertex*)vertices vSize:(size_t)vsize indices:(const GLushort*)indices iSize:(size_t)isize objectName:(NSString*)objectName{
-    _vaoID = [LoaderHelper loadToVBOS:vertices verticesSize:vsize indices:indices indicesSize:isize objectName:objectName];
+    _vaoID = [LoaderHelper loadToVBOS:vertices verticesSize:(int)vsize indices:indices indicesSize:(int)isize objectName:objectName];
     
-    _numVertices = isize/sizeof(GLushort);
+    _numVertices = (int)isize/(int)sizeof(GLushort);
 }
 -(void)loadToTexture:(NSString*)fileName{
     _texture = [LoaderHelper loadTexture:fileName];
